rename rx -> sx, cleanup
