+++ src/components/Card.tsx
@@ -1,4 +1,4 @@
-import { themed } from "../styleComposition"
+import { themed } from "../theming"
 import { RosesSC } from "../types"
 
 export const Card: RosesSC = themed("Card")
