+++ src/components/Typography.stories.js
@@ -9,19 +9,19 @@ storiesOf("Typography", module)
   .add("Headings", () => (
     <Box>
       <Heading>Heading 1 (default: fontSize=5)</Heading>
-      <Heading rx={{ color: "primary", fontSize: 4 }}>
+      <Heading sx={{ color: "primary", fontSize: 4 }}>
         Heading 2 (fontSize=4)
       </Heading>
-      <Heading rx={{ color: "secondary", fontSize: 3 }}>Heading 3</Heading>
-      <Heading rx={{ fontSize: 2 }}>Heading 4</Heading>
-      <Heading rx={{ fontSize: 1 }}>Heading 5</Heading>
-      <Heading rx={{ fontSize: 0 }}>Heading 6 (fontSize=0)</Heading>
+      <Heading sx={{ color: "secondary", fontSize: 3 }}>Heading 3</Heading>
+      <Heading sx={{ fontSize: 2 }}>Heading 4</Heading>
+      <Heading sx={{ fontSize: 1 }}>Heading 5</Heading>
+      <Heading sx={{ fontSize: 0 }}>Heading 6 (fontSize=0)</Heading>
     </Box>
   ))
   .add("Fonts", () => (
-    <Box rx={{ mx: "auto", width: "600px" }}>
-      <Heading rx={{ my: 4, fontSize: 3, fontFamily: "sans" }}>Sans</Heading>
-      <Text rx={{ fontFamily: "sans" }}>
+    <Box sx={{ mx: "auto", width: "600px" }}>
+      <Heading sx={{ my: 4, fontSize: 3, fontFamily: "sans" }}>Sans</Heading>
+      <Text sx={{ fontFamily: "sans" }}>
         Minus consequuntur eligendi. Aliquid veritatis dolor ipsa amet. Nisi
         provident consequatur. Vitae et amet qui aut. Ullam rerum rerum illum
         sit ut cumque magni magni sed. Voluptatum ut autem.
@@ -33,8 +33,8 @@ storiesOf("Typography", module)
         Laborum dignissimos et consequatur. Id et voluptatem enim qui voluptatem
         sint. Qui aut nam quisquam facilis illo corrupti nostrum.
       </Text>
-      <Heading rx={{ my: 4, fontSize: 3, fontFamily: "serif" }}>Serif</Heading>
-      <Text rx={{ fontFamily: "serif" }}>
+      <Heading sx={{ my: 4, fontSize: 3, fontFamily: "serif" }}>Serif</Heading>
+      <Text sx={{ fontFamily: "serif" }}>
         Consequatur minima sint eligendi ut et quae odit commodi. Harum veniam
         aut quaerat velit enim. Sint non consequatur rerum possimus blanditiis
         debitis incidunt est.
@@ -48,9 +48,9 @@ storiesOf("Typography", module)
         error saepe. Vel reprehenderit eum voluptas sint dicta culpa qui quam
         rerum. Reiciendis facilis aliquid.
       </Text>
-      <Heading rx={{ my: 4, fontSize: 3, fontFamily: "monospace" }}>
+      <Heading sx={{ my: 4, fontSize: 3, fontFamily: "monospace" }}>
         Monospace
       </Heading>
-      <Text rx={{ fontFamily: "monospace" }}>Qui odio nulla.</Text>
+      <Text sx={{ fontFamily: "monospace" }}>Qui odio nulla.</Text>
     </Box>
   ))
