+++ src/theming.test.tsx
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import { jsx } from "@emotion/core"
-import { withStyleProps, themed } from "./styleComposition"
+import { themed } from "./theming"
 import { RosesTheme } from "./Theme"
 import renderer from "react-test-renderer"
 import { baseTheme } from "./Theme/baseTheme"
@@ -13,22 +13,11 @@ const theme = {
     ...baseColors,
     primary: "red",
     secondary: "green",
+    muted: "tomato",
   },
 }
 
 expect.extend(matchers)
-
-const render = (contents: React.ReactNode) =>
-  renderer.create(<RosesTheme theme={theme}>{contents}</RosesTheme>)
-
-describe("withStyleProps", () => {
-  it("adds an rx prop that takes styled-system/css inputs", () => {
-    const Widget = withStyleProps("div")
-    const tree = render(<Widget rx={{ bg: "secondary" }} />)
-    expect(tree.toJSON()).toMatchSnapshot()
-    // expect(tree).toHaveStyleRule("background", "brown") //, { target: `${Widget}` })
-  })
-})
 
 describe("themed()", () => {
   const localTheme = {
@@ -60,5 +49,15 @@ describe("themed()", () => {
       </RosesTheme>
     )
     expect(tree2.toJSON()).toMatchSnapshot()
+  })
+
+  it("has an sx prop for theme-aware styles", () => {
+    const DivWidget = themed("Widget")
+    const tree = renderer.create(
+      <RosesTheme theme={localTheme}>
+        <DivWidget sx={{ color: "muted" }} />
+      </RosesTheme>
+    )
+    expect(tree.toJSON()).toMatchSnapshot()
   })
 })
