+++ src/types.ts
@@ -13,5 +13,5 @@ export type RosesStyledComponent = RosesSC
 export interface RosesStyleProps {
   variant?: string
   /** A SystemStyleObject - theme-aware css to be applied last (see @styled-system/css) */
-  rx?: SystemStyleObject
+  sx?: SystemStyleObject
 }
